Service for Topcoder API v2 and v3. Proper authentication handling.
NOTE: To be successfully authenticated you should be logged in the
Topcoder accounts-app, and run the code at local.topcoder-dev.com:3000
for development backend, or local.topcoder.com for production backend.
If you run from a wrong domain, the code will fail to get auth token for
Topcoder API v2 even if you are properly logged in. In such case, if you
are using development build of the code, you'll see an error message in
console.