Unit tests clean-up
Clean up of unit tests.
In this commit:

1. Test coverage is incremented to:
- branches:	67.75%
- functions: 	83.00%
- lines: 	83.50%
- statements: 	82.75%

2. No unit tests fail.

3. All unit tests written in a good way (sure, there is a room for
improvement, but in absense of better ideas, the pattern of existing
tests can be followed as it is).

DISCLAIMER: Some unit tests still may fail randomly sometimes. The
reason is that when they are executed in random order they may change
process.env (environment variables) in a way which affect other tests.
If a test fails, most probably you want to:

- delete process.env.FRONT_END - this will ensure that code assumes
server-side execution, and configuration is correctly loaded.

- process.env.NODE_ENV = 'development' - this will ensure that
development configuration is loaded.

Both are necessary for consistent rendering of React components.
Probably, rather than adding these checks to all tests, we should just
ensure that those few tests that touch these variables, set them to
proper defaults in the end. To be investigated.