Fix: Return URL in Login and Register buttons
Fixes #74.

Return URL for Login and Regiser buttons in header and footer of TC
Community mini-sites was hard-coded to lead to Wipro community. Now it
always redirects back to the page where the button was clicked. As a
minor side-effect, the buttons were converted from <a> links into
<button> elements. Unfortunately, it prevents them working with
right-mouse-click menu, but prevents any inconsistent behavior across
our dev/prod setups, or problems with the server-side rendering.