Challenge listing is updated to load all active challenges at once
Currently, backend allows to load at most 50 challenge objects by a
single request. In prod we have typically around 30-50 active challenge
at the same time, thus most of the time this change won't have any
impact on the performance. Even if, eventually, there will be more than
50 active challenges, a few additional requests are not a big deal for
the peformance.

On the other hand, assuming that all active challenges are already
loaded into the frontend allows to simplify a lot of code at the
frontend side, and to provide a better and more consistent user
experience when filtering&counting challenges in the challenge listing.