Fix: External links don't work
Fixes #99

NOTE: this commit itself provides a general solution to the problem, and
fixes two links explicitely mentioned in the original issue ticket. To
fix the problem in any other place:
1) Find the <Link> or <NavLink> element corresponding to the link;
2) Update its import statement to import these elements from
`utils/router` rather than from `react-router-dom`.
3) Update unit tests: our customized <Link> and <NavLink> elements need
access to the Redux store. Be sure that in your tests their containers
are wrapped with <Provider store={store}>...</Provider>, where `store`
object can be as simple as:
```
{
dispatch: () => _.noop,
getState: () => { hostname: 'HOSTNAME APPROPRITATE FOR THE TEST' },
subscribe: _.noop,
}
```

NOTE: To decide whether an URL leads outside of the app or not, our
custom <Link> and <NavLink> components compare hostname from the URL
with the hostname of the page, where these components are rendered. It
is possible to run into situation, when this does not work right,
because different endpoints in the same domain are served from different
apps. In this case a different solution will be necessary.