Refactoring: Community listings
Currently here are three places where TC communities, visible to a
vistor, are listed: (1) community filters in the dashboard; (2)
community switch in challenge listings; (3) community switch in the TC
community header.

This commit ensures that we do not dublicate the data necessary for (1)
and (2) in different places of Redux store, and that we use the same
logic and code to decide which communities are visible to user.

(3) is not yet connected to this code, and still uses lists of
communities hard-coded into each community config. This will be changed
in one of the next commits.

Fixes #100