Fix: Slicing of challenges visible in the Dashboard
Before only 8 active challenges were shown in the dashboard, now this
restriction is lifted.

Even, if we want to bring the restriction back, mind that there was a
problem fixed by this change: the array of user's challenges was sliced
before passing it to community filters in the dashboard (to work
properly, we are supposed first to filter challenges by communites, and
only then restrict the number of challenges visible in the dashboard).