Removes CSS optimization on development Webpack builds
It significantly speeds-up HMR. Trade-off is that browser's dev tools
become extremely slow, due to large size of CSS code (~5mb) (it is a way
larger than in prod, because generated classnames are longer, and some
dublicates of the code are not cleaned out). However, right now ~4.5mb
of CSS is the size of the main CSS bundle. In the next commits, the new
approach for code-splitting will be implemented on all existing routes,
and the size of the main CSS bundle will become a way smaller.
Presumably, even in dev, without CSS optimizations, regular pages won't
have more than ~1mb of CSS code; thus, hopefully, browser's dev tools
will work fast again.