A better fix of extract-css-chunks-webpack-plugin problem
The code of plugin has been updated to accept "justExtract" option,
which tells the plugin just to extract CSS and to avoid doing any other
staff we do not need.

Temporarly serving updated plugin from my own fork, will be switched to
the npm version as soon as the change is merged in and published.