Refactoring of communities-related code - Step I
Prior to this point, most of the community-related logic (loading of
meta-data, auth checks, loading of other data, and connection of all
community components to Redux) was handled by
/src/shared/containers/tc-communities/Page/index.jsx. It was a big
module with a messy logic and some performance issues, which had to be
refactored into something more maintainable.

In this commit the meta-data loading and auth checks logic have been
split into a separate Loader container, and it is ensured that it works
optimally (does not reload any data when not-necessary).