Mock for challenge terms service
Adds a mock implementation of challenge terms service (both for simple
Topcoder terms, and for DocuSign flow). If enabled in config, the app
will use it, allowing to test challenge terms flow in a convenient way,
and to carry on related development.

It also can be used to mock terms service in Jest unit tests.