Fix of react-css-themr example
It turns out that our setup does not guarantee that order of styles will
match the order of imports (numerous plugins may shake order of CSS, it
is something a way too tricky to solve). Generally, it is not a big deal
for us, but it is a problem, when react-css-themr deeply merges themes
from different sources. It is necessary to be very careful, and use
!important rule for the styles that should take priority.