Replaces react-css-themr by react-css-super-themr + Fixes a communities bug
Fixed communities bug: before, when we navigated from one community to
another, and some metadata fields were defined in the first community,
but no in the other, they had not been reseted to undefined. Fixed by
ensuring that proper defaults are set for community metadata on
server-side instead of undefined.

Fixes #460