Update of Modal component: in general, and for challenge terms
Updates layout (mostly, sizing behavior in desktop and mobile), and
behavior (blocking of the parent window's scrolling) of the standard
Modal component in general; and in the case of challenge terms, in
particular. In the later case, the modal's content is updated so that
most of the content is scrollable, thus improving user's experience.

+ Updated preview URL in Submission Management Page