Fix: Optimization of challenge listing (re-)rendering
It was overlooked that FilterPanel and ChallengeListing containers were
generating several object-like props inside their mapStateToProps(..)
functions. Because of this, ReactJS though that props are updated each
time he evaluated such possibility; which hurt performance in some cases
(e.g. renaming of custom filters in the sidebar was really suffering
from this issue).