Fix (hopefully) for the community registration flow
It turned out that community registration flow did not work for
communities protected by DocuSign terms, if at a page there were
multiple instances of Join Community buttons. The reason was that
multiple terms modals opened simultaneously (and invisibly to the
end-user), and failed when they tried to load the same DocuSign term.